Add support for MaterializeCSS material box
The MaterializeCSS library comes with support for rendering images as thumbnails
that expand to full-screen on click. It requires running an initialization code
on our default layout page, which has been added here.

In addition, we add a style for .materialboxed.initialized.active that allow
the material box to become full screen even when the image is part of a column.

Read more about material box: http://materializecss.com/media.html